Fix speed_test, use long type for the cycle counter

+++ examples/speed_test/speed_test.ino
@@ -59,7 +59,7 @@ void loop()
     // Static variables.
     //
     static long lastMillis = 0;
-    static int cycles = 0;
+    static long cycles = 0;
 
     long currentMillis = millis();
